feat: move metadata from frontend
