+++ bin/faucet/src/api/frontend.rs
@@ -1,20 +1,6 @@
-use axum::{
-    Json,
-    extract::State,
-    response::{Html, IntoResponse, Response},
-};
+use axum::response::{Html, IntoResponse, Response};
 use axum_extra::response::{Css, JavaScript};
 use http::header::{self};
-use miden_faucet_client::{FaucetId, types::AssetOptions};
-
-/// Describes the faucet metadata.
-///
-/// More specifically, the faucet's account ID and allowed mint amounts.
-#[derive(Clone, serde::Serialize)]
-pub struct Metadata {
-    pub id: FaucetId,
-    pub asset_amount_options: AssetOptions,
-}
 
 pub async fn get_index_html() -> Html<&'static str> {
     Html(include_str!("../static/index.html"))
@@ -42,8 +28,4 @@ pub async fn get_favicon() -> Response {
         include_bytes!("../static/favicon.ico"),
     )
         .into_response()
-}
-
-pub async fn get_metadata(State(metadata): State<&'static Metadata>) -> Json<&'static Metadata> {
-    Json(metadata)
 }
